+++ gwcs/wcs.py
@@ -15,7 +15,6 @@ __all__ = ['WCS']
 
 
 class WCS:
-
     """
     Basic WCS class.
 
@@ -260,7 +259,7 @@ class WCS:
 
     def invert(self, *args, **kwargs):
         """
-        Invert coordnates.
+        Invert coordinates.
 
         The analytical inverse of the forward transform is used, if available.
         If not an iterative method is used.
@@ -278,7 +277,9 @@ class WCS:
             Output value for inputs outside the bounding_box (default is np.nan).
         """
         if not utils.isnumerical(args[0]):
-            args = utils._get_values(self.unit, *args)
+            args = self.output_frame.coordinate_to_quantity(*args)
+            if not self.forward_transform.uses_quantity:
+                args = utils.get_values(self.output_frame.unit, *args)
 
         output = kwargs.pop('output', None)
         if 'with_bounding_box' not in kwargs:
@@ -291,7 +292,7 @@ class WCS:
         except (NotImplementedError, KeyError):
             result = self._invert(*args, **kwargs)
 
-        if output == 'numericals_plus':
+        if output == 'numericals_plus' and self.input_frame:
             if self.input_frame.naxes == 1:
                 return self.input_frame.coordinates(result)
             else:
@@ -329,7 +330,10 @@ class WCS:
         """
         transform = self.get_transform(from_frame, to_frame)
         if not utils.isnumerical(args[0]):
-            args = utils._get_values(self.unit, *args)
+            inp_frame = getattr(self, from_frame)
+            args = inp_frame.coordinate_to_quantity(*args)
+            if not transform.uses_quantity:
+                args = utils.get_values(inp_frame.unit, *args)
 
         output = kwargs.pop("output", None)
         if 'with_bounding_box' not in kwargs: